Remove (now) unnecessary casting from GCD implementation.
Monotone-Parent: 878032db68fd1abcfaa57a1bdb86fb6a16c571fb
Monotone-Revision: f6f2b4da754790bd204a9b64684d7cf1a599f146

+++ lib/symbolic/rational.cpp
@@ -168,9 +168,9 @@ void Rational::normalise()
     denominator = -denominator;
   }
 
-  const unsigned long factor = gcd(std::abs(numerator), std::abs(denominator));
-  numerator /= static_cast<long>(factor);
-  denominator /= static_cast<long>(factor);
+  const long factor = gcd(std::abs(numerator), std::abs(denominator));
+  numerator /= factor;
+  denominator /= factor;
 
   assert(denominator != 0);
 }
